Implement requested refacotring for follower indices
